minor fix: removed some gcc warnings and make test.sh python3-ready.

+++ cpp/include/tapas/common.h
@@ -91,7 +91,7 @@ namespace tapas {
 using std::string;
 using std::ostream;
 
-typedef long index_t;
+typedef size_t index_t;
 
 inline void Exit(int status, const char *file, const char *func, int line) {
   if (status) {
